# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserRegisterInputRole = typeof UserRegisterInputRole[keyof typeof UserRegisterInputRole];


export const UserRegisterInputRole = {
  passenger: 'passenger',
  driver: 'driver',
} as const;

export interface UserRegisterInput {
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  password: string;
  role: UserRegisterInputRole;
}

export interface UserLoginInput {
  phone: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  passenger: 'passenger',
  driver: 'driver',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  role: UserRole;
  isActive: boolean;
  isOnline: boolean;
  /** @nullable */
  currentLat?: number | null;
  /** @nullable */
  currentLng?: number | null;
  rating: number;
  ratingCount: number;
  createdAt: string;
}

export interface AuthResult {
  token: string;
  user: User;
}

export interface UserStatusUpdate {
  isActive: boolean;
}

export interface DriverStatusUpdate {
  isOnline: boolean;
}

export interface LocationUpdate {
  lat: number;
  lng: number;
}

export interface DriverLocation {
  id: number;
  name: string;
  lat: number;
  lng: number;
  rating: number;
  /** @nullable */
  vehicleType: string | null;
}

export type VehicleVehicleType = typeof VehicleVehicleType[keyof typeof VehicleVehicleType];


export const VehicleVehicleType = {
  economico: 'economico',
  ejecutivo: 'ejecutivo',
  taxi: 'taxi',
  moto: 'moto',
} as const;

export interface Vehicle {
  id: number;
  driverId: number;
  plate: string;
  model: string;
  brand: string;
  color: string;
  vehicleType: VehicleVehicleType;
  createdAt: string;
}

export type VehicleInputVehicleType = typeof VehicleInputVehicleType[keyof typeof VehicleInputVehicleType];


export const VehicleInputVehicleType = {
  economico: 'economico',
  ejecutivo: 'ejecutivo',
  taxi: 'taxi',
  moto: 'moto',
} as const;

export interface VehicleInput {
  plate: string;
  model: string;
  brand: string;
  color: string;
  vehicleType: VehicleInputVehicleType;
}

export type TripStatus = typeof TripStatus[keyof typeof TripStatus];


export const TripStatus = {
  pending: 'pending',
  accepted: 'accepted',
  driver_arriving: 'driver_arriving',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type TripPaymentMethod = typeof TripPaymentMethod[keyof typeof TripPaymentMethod];


export const TripPaymentMethod = {
  cash: 'cash',
  card: 'card',
} as const;

/**
 * @nullable
 */
export type TripDriver = {
  id?: number;
  name?: string;
  phone?: string;
  rating?: number;
  /** @nullable */
  currentLat?: number | null;
  /** @nullable */
  currentLng?: number | null;
} | null;

export interface Trip {
  id: number;
  passengerId: number;
  /** @nullable */
  driverId?: number | null;
  status: TripStatus;
  originLat: number;
  originLng: number;
  originAddress: string;
  destinationLat: number;
  destinationLng: number;
  destinationAddress: string;
  vehicleType: string;
  estimatedPrice: number;
  /** @nullable */
  finalPrice?: number | null;
  /** @nullable */
  distanceKm?: number | null;
  paymentMethod: TripPaymentMethod;
  /** @nullable */
  cancelReason?: string | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  cancelledAt?: string | null;
  createdAt: string;
  passenger?: User;
  /** @nullable */
  driver?: TripDriver;
}

export type TripInputVehicleType = typeof TripInputVehicleType[keyof typeof TripInputVehicleType];


export const TripInputVehicleType = {
  economico: 'economico',
  ejecutivo: 'ejecutivo',
  taxi: 'taxi',
  moto: 'moto',
} as const;

export type TripInputPaymentMethod = typeof TripInputPaymentMethod[keyof typeof TripInputPaymentMethod];


export const TripInputPaymentMethod = {
  cash: 'cash',
  card: 'card',
} as const;

export interface TripInput {
  originLat: number;
  originLng: number;
  originAddress: string;
  destinationLat: number;
  destinationLng: number;
  destinationAddress: string;
  vehicleType: TripInputVehicleType;
  paymentMethod: TripInputPaymentMethod;
  estimatedPrice?: number;
}

export type TripStatusUpdateStatus = typeof TripStatusUpdateStatus[keyof typeof TripStatusUpdateStatus];


export const TripStatusUpdateStatus = {
  accepted: 'accepted',
  driver_arriving: 'driver_arriving',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface TripStatusUpdate {
  status: TripStatusUpdateStatus;
  /** @nullable */
  cancelReason?: string | null;
  /** @nullable */
  finalPrice?: number | null;
}

export interface Message {
  id: number;
  tripId: number;
  senderId: number;
  content: string;
  senderName?: string;
  createdAt: string;
}

export interface MessageInput {
  content: string;
}

export interface Rating {
  id: number;
  tripId: number;
  raterId: number;
  rateeId: number;
  score: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface RatingInput {
  score: number;
  /** @nullable */
  comment?: string | null;
}

export interface PanicAlert {
  driverId: number;
  driverName: string;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  timestamp: string;
  message: string;
  notifiedDrivers: number;
}

export interface AdminStats {
  totalUsers: number;
  totalDrivers: number;
  totalPassengers: number;
  activeTrips: number;
  completedTrips: number;
  totalRevenue: number;
  onlineDrivers: number;
  pendingTrips?: number;
}

export type ListTripsParams = {
status?: string;
limit?: number;
offset?: number;
};

export type GetNearbyTripsParams = {
lat: number;
lng: number;
radius?: number;
};

export type GetNearbyDriversParams = {
lat: number;
lng: number;
vehicleType?: string;
};

export type ListAdminUsersParams = {
role?: string;
isActive?: boolean;
limit?: number;
offset?: number;
};

export type ListAdminTripsParams = {
status?: string;
limit?: number;
offset?: number;
};